requirements.txt
fastapi
uvicorn
celery
redis
sqlalchemy
psycopg2-binary
flower
python-dotenv
